Remake structure on modular
