+++ src/about-us/about-us-index.js
@@ -1,51 +1,10 @@
 import "../index.scss";
-window.addEventListener('DOMContentLoaded', ()=> {
-  // fix transition on page load
-  function returnTransition() {
-    document.querySelector("body").classList.remove("no-transition");
-  }
-   window.onload = () => {
-     returnTransition();
-   };
-  //atom animation
-  const circles = document.querySelectorAll("#circle");
-  const atom = document.querySelector("#atom");
+import transitionFix from "../modules/transitionFix";
+import animationLogo from "../modules/animationLogo";
+import menu from "../modules/menu";
 
-  circles.forEach((circle) => {
-    atom.addEventListener("mouseover", () => {
-      circle.setAttribute("cx", "0");
-      circle.setAttribute("cy", "0");
-    });
-  });
-
-
-  const navbar = document.querySelector(".navbar");
-  const burger = document.querySelector(".burger");
-  const blur = document.querySelector(".wrapper-blur");
-
-  //burger menu
-  function toggleMenu() {
-    navbar.classList.toggle("navbar--open");
-    burger.classList.toggle("burger--open");
-    blur.classList.toggle("wrapper-blur--show");
-    document.body.style.overflow === "" ? document.body.style.overflow = "hidden" : document.body.style.overflow = "";
-  }
-
-  burger.addEventListener("click", toggleMenu);
-  blur.addEventListener('click', ()=>{
-    if(blur.classList.contains('wrapper-blur--show')){
-      toggleMenu();
-    }
-  });
-
-  // fix menu on resizing window
-  let timeOutFunctionId;
-  window.addEventListener('resize', ()=>{
-    document.body.classList.add("no-transition");
-    if (window.matchMedia('(min-width: 768px)').matches && navbar.classList.contains('navbar--open')){
-      toggleMenu();
-    }
-    clearTimeout(timeOutFunctionId);
-    timeOutFunctionId = setTimeout(returnTransition, 250);
-  });
+window.addEventListener("DOMContentLoaded", () => {
+  transitionFix();
+  animationLogo();
+  menu();
 });
