+++ src/about-us/about-us.html
@@ -6,7 +6,6 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Dive Into</title>
-  <link rel="stylesheet" type="text/css" href="style.css" />
 </head>
 
 <body class="no-transition">
