+++ src/members/members.html
@@ -160,9 +160,9 @@
         <div class="members__main">
           <ul class="section__list members__list">
             <li class="section__blank hide">
-              <img src="/assets/img/search.svg" alt="search">
-              <p>Sorry! We couldn't find any members based on your filters</p>
-              </li>
+              <img class="section__blank-item" src="/assets/img/search.svg" alt="search">
+              <p class="section__blank-item">Sorry! We couldn't find any members based on your filters</p>
+            </li>
           </ul>
           <div class="section__filters members__filters">
             <svg data-close class="section__filters-close" width="30" height="30" viewBox="0 0 30 30" fill="none"
