+++ src/modules/animationLogo.js
@@ -1,0 +1,13 @@
+function animationLogo () {
+  //atom animation
+  const circles = document.querySelectorAll("#circle");
+  const atom = document.querySelector("#atom");
+
+  circles.forEach((circle) => {
+    atom.addEventListener("mouseover", () => {
+      circle.setAttribute("cx", "0");
+      circle.setAttribute("cy", "0");
+    });
+  });
+}
+export default animationLogo;
