+++ src/modules/getResource.js
@@ -1,0 +1,10 @@
+async function getResource(url) {
+  let res = await fetch(url);
+
+  if (!res.ok) {
+    throw new Error(`Could not fetch ${url}, status: ${res.status}`);
+  }
+  return await res.json();
+}
+
+export default getResource;
